로그아웃 후 재로그인시 이전정보 가져오는 문제 수정
