+++ app/src/main/java/com/wannagohome/lens_review_android/support/AccessKeyHelper.kt
@@ -2,15 +2,19 @@ package com.wannagohome.lens_review_android.support
 
 import android.content.Context
 import com.wannagohome.lens_review_android.AppComponents
+import com.wannagohome.lens_review_android.network.InterceptorManager
+import org.koin.core.KoinComponent
+import org.koin.core.inject
 
-object AccessKeyHelper {
+object AccessKeyHelper : KoinComponent {
     private const val KEY_TOKEN = "token"
     private const val KEY_AUTH = "auth"
+    private val interceptorManager: InterceptorManager by inject()
 
     fun addToken(token: String) {
         val editor = AppComponents.applicationContext.getSharedPreferences(KEY_AUTH, Context.MODE_PRIVATE).edit()
         editor.putString(KEY_TOKEN, token)
-        editor.apply()
+        editor.commit()
     }
 
     fun readToken(): String {
@@ -26,11 +30,13 @@ object AccessKeyHelper {
         val editor = AppComponents.applicationContext.getSharedPreferences(KEY_AUTH, Context.MODE_PRIVATE).edit()
         editor.remove(KEY_TOKEN)
         editor.apply()
+        interceptorManager.clearAuth()
     }
 
     fun deleteTokenSync() {
         val editor = AppComponents.applicationContext.getSharedPreferences(KEY_AUTH, Context.MODE_PRIVATE).edit()
         editor.remove(KEY_TOKEN)
         editor.commit()
+        interceptorManager.clearAuth()
     }
 }
