+++ app/src/main/java/com/wannagohome/lens_review_android/support/di/NetworkModule.kt
@@ -1,6 +1,7 @@
 package com.wannagohome.lens_review_android.support.di
 
 import com.wannagohome.lens_review_android.network.AppRetrofitBuilder
+import com.wannagohome.lens_review_android.network.InterceptorManager
 import com.wannagohome.lens_review_android.network.NetworkConfig
 import com.wannagohome.lens_review_android.network.lensapi.LensApiClient
 import com.wannagohome.lens_review_android.network.lensapi.LensApiInterface
@@ -12,7 +13,10 @@ val networkModule = module {
         LensApiClient(
             AppRetrofitBuilder(NetworkConfig.API_BASE_URL)
                 .build()
+
                 .create(LensApiInterface::class.java)
         )
     }
+
+    single { InterceptorManager() }
 }
